Create PizzaParty
This file was from a homework assignment intended to help pizza party planners. It takes the number of pizzas bought, the number of slices per pizza, and the number of guests, divided into adults and children. It then outputs the amount of slices each child will get (they are equally distributed what's left after each adult gets 2 slices of pizza), along with the remainder of pizza slices.