Create Asphalt
For this project, a scenario was given to me to help a construction team determine how many truckloads of asphalt to buy in order to pave a new road, and the total cost of the required asphalt. I was tasked with getting input to calculate the volume of the asphalt the road need according to its size, and the information I was given was that a road lane was 12 feet wide, asphalt weighs 145 pounds per cubic foot, a truckload of asphalt can carry 5 tons (equal to 10,000 pounds), and asphalt costs $150 per ton (so one truckload of asphalt costs $750). This road assumes no bike lane or shoulder.